refactor: 구현코드 main 으로 이동
