+++ src/main/kotlin/blackjack/Symbol.kt
@@ -1,0 +1,8 @@
+package blackjack
+
+enum class Symbol(private val symbolName: String) {
+    HEARTS("하트"),
+    CLUBS("클로버"),
+    SPADES("스페이드"),
+    DIAMONDS("다이아몬드")
+}
