+++ src/test/kotlin/blackjack/CardTest.kt
@@ -52,55 +52,4 @@ class CardTest {
             Card("J", Symbol.DIAMONDS)
         )
     }
-
-    private fun blackjack(initializer: BlackjackBuilder.() -> Unit): BlackjackBuilder {
-        return BlackjackBuilder().apply(initializer)
-    }
-
-    class BlackjackBuilder {
-        var deck: List<Card> = emptyList()
-
-        fun normal(numbers: IntRange) {
-            deck = deck + numbers.map { toSymbolCard(it.toString()) }.flatten()
-        }
-
-        fun ace() {
-            deck = deck + toSymbolCard("A")
-        }
-
-        fun jack() {
-            deck = deck + toSymbolCard("J")
-        }
-
-        fun queen() {
-            deck = deck + toSymbolCard("Q")
-        }
-
-        fun king() {
-            deck = deck + toSymbolCard("K")
-        }
-
-        private fun toSymbolCard(name: String): List<Card> {
-            return Symbol.values().map { Card(name, it) }
-        }
-
-        fun build(): Blackjack = Blackjack(deck)
-    }
-
-    class Blackjack(val deck: List<Card>)
-
-    data class Card(private val name: String, private val symbol: Symbol) {
-        val number: List<Int> = when (name) {
-            "A" -> listOf(1, 11)
-            "J", "Q", "K" -> listOf(10)
-            else -> listOf(name.toInt())
-        }
-    }
-
-    enum class Symbol(private val symbolName: String) {
-        HEARTS("하트"),
-        CLUBS("클로버"),
-        SPADES("스페이드"),
-        DIAMONDS("다이아몬드")
-    }
 }
