+++ src/test/kotlin/blackjack/PlayerTest.kt
@@ -1,10 +1,7 @@
 package blackjack
 
-import blackjack.CardTest.Card
-import blackjack.CardTest.Symbol
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
-import kotlin.math.abs
 
 class PlayerTest {
     @Test
@@ -63,64 +60,5 @@ class PlayerTest {
                     draw(Card("8", Symbol.CLUBS))
                 }.score()
         ).isEqualTo(24)
-    }
-
-    interface Player {
-        val name: String
-        val cards: List<Card>
-        fun score(): Int
-        fun draw(card: Card)
-        fun draw(
-            nextCard: () -> Card,
-            isDraw: (name: String) -> Boolean = { true }
-        )
-
-        class Person(override val name: String) : Player {
-            private var _cards: List<Card> = emptyList()
-            override val cards: List<Card>
-                get() = _cards
-
-            override fun score(): Int {
-                return _cards.fold(listOf(0)) { acc, card ->
-                    acc.flatMap { score -> card.number.map { it + score } }
-                }.closeTo(21)
-            }
-
-            override fun draw(card: Card) {
-                _cards = _cards + card
-            }
-
-            override fun draw(nextCard: () -> Card, isDraw: (name: String) -> Boolean) {
-                if (isDraw(name)) {
-                    draw(nextCard())
-                }
-            }
-
-            private fun List<Int>.closeTo(number: Int): Int {
-                val sorted = map { it to abs(it - number) }
-                    .sortedBy { it.second }
-                    .map { it.first }
-                val result = sorted.firstOrNull { it <= 21 }
-                return result ?: sorted.firstOrNull() ?: 0
-            }
-
-            override fun equals(other: Any?): Boolean {
-                if (this === other) return true
-                if (javaClass != other?.javaClass) return false
-
-                other as Person
-
-                if (name != other.name) return false
-                if (_cards != other._cards) return false
-
-                return true
-            }
-
-            override fun hashCode(): Int {
-                var result = name.hashCode()
-                result = 31 * result + _cards.hashCode()
-                return result
-            }
-        }
     }
 }
